Add Unique URLs for IndividualResourcePage
