+++ frontend/src/pages/IndividualResourcePage/IndividualResourcePage.js
@@ -1,46 +1,11 @@
-import React, { useEffect, useState } from 'react'
-import {
-    Banner,
-    IndividualResourceTileGroup,
-    CategoryButtonGroup,
-    TextBlock,
-} from '../../components/components.js'
+import React, { useEffect } from 'react'
 import { useLocation } from 'react-router-dom'
-import URL_PATH from '../../links.js'
+import { ResourcePopUp } from '../../components/components.js'
 
 function IndividualResourcePage({ setActiveLink }) {
     const location = useLocation()
-    const { individualIDs, title, description, imageUrl } = location.state || {}
-
-    const individualIDsQueryParam = new URLSearchParams({
-        listOfResourceIDs: JSON.stringify(individualIDs),
-    }).toString()
-
-    const [resourceMapper, setResourceMapper] = useState({})
-    const [resourceList, setResourceList] = useState([])
-
-    useEffect(() => {
-        // URL_PATH imported from frontend/src/links.js
-        // combined with subdirectory to make the full URL
-        const subdirectory = `/resources/individualResources?${individualIDsQueryParam}`
-        fetch(URL_PATH + subdirectory)
-            .then((response) => response.json())
-            .then((json) => {
-                let tempDict = {}
-                let tempList = []
-
-                json.forEach((resource) => {
-                    tempDict[resource.Category] = resource
-                    tempList.push(resource)
-                })
-
-                setResourceMapper(tempDict)
-                setResourceList(tempList)
-
-                window.scrollTo(0, 0)
-            })
-            .catch((error) => console.error(error))
-    }, [])
+    const pathSegments = location.pathname.split('/')
+    const id = pathSegments[pathSegments.length - 1]
 
     useEffect(() => {
         setActiveLink('/Resources')
@@ -48,28 +13,7 @@ function IndividualResourcePage({ setActiveLink }) {
 
     return (
         <div>
-            <Banner imageUrl={imageUrl} />
-
-            <CategoryButtonGroup
-                title={title}
-                names={Object.keys(resourceMapper)}
-                locations={Object.keys(resourceMapper)}
-            />
-
-            <TextBlock text={description} />
-
-            {Object.keys(resourceMapper).map((categoryName) => {
-                let result = resourceList.filter(
-                    (resource) => categoryName === resource.Category,
-                )
-                return (
-                    <IndividualResourceTileGroup
-                        id={categoryName}
-                        title={categoryName}
-                        resources={result}
-                    />
-                )
-            })}
+            <ResourcePopUp id={id} />
         </div>
     )
 }
